chore: Refactor directory structure
Move containers into "pods" (feature-based directories): Authentication, Contacts, and Stories.
