+++ app/containers/Contacts/ContactCard/EditContact.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Field, reduxForm } from 'redux-form/immutable';
+import { Field, reduxForm } from '../../../../node_modules/redux-form/immutable';
 import { Button, Glyphicon } from 'react-bootstrap';
 import styles from './styles.css';
 import {
