+++ app/containers/Contacts/ContactCard/default-avatar.png
