+++ app/containers/Contacts/ContactCard/reducer.js
