+++ app/containers/Stories/reducer.js
